+++ heylogs-api/src/main/java/nbbrd/heylogs/Resource.java
@@ -1,0 +1,10 @@
+package nbbrd.heylogs;
+
+import lombok.NonNull;
+
+@lombok.Value
+public class Resource {
+
+    @NonNull String type;
+    @NonNull String id;
+}