+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/Format.java
@@ -2,6 +2,7 @@ package nbbrd.heylogs.spi;
 
 import lombok.NonNull;
 import nbbrd.heylogs.Failure;
+import nbbrd.heylogs.Resource;
 import nbbrd.heylogs.Status;
 import nbbrd.service.Quantifier;
 import nbbrd.service.ServiceDefinition;
@@ -22,4 +23,6 @@ public interface Format {
     void formatFailures(@NonNull Appendable appendable, @NonNull String source, @NonNull List<Failure> failures) throws IOException;
 
     void formatStatus(@NonNull Appendable appendable, @NonNull String source, @NonNull Status status) throws IOException;
+
+    void formatResources(@NonNull Appendable appendable, @NonNull List<Resource> resources) throws IOException;
 }