+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add check on GitHub Pull Request links [#173](https://github.com/nbbrd/heylogs/issues/173)
 - Add check on GitHub mentions of people and teams [#157](https://github.com/nbbrd/heylogs/issues/157)
 - Add check on GitHub commit SHAs [#223](https://github.com/nbbrd/heylogs/issues/223)
+- Improve list command output [#231](https://github.com/nbbrd/heylogs/issues/231)
 
 ## [0.7.2] - 2023-11-10